isolated inspect_all test
