+++ config_defaults/subtests/docker_cli/dockerinspect.ini
@@ -1,7 +1,7 @@
 [docker_cli/dockerinspect]
 docker_timeout = 120.0
 remove_after_test = yes
-subsubtests = inspect_container_simple,inspect_all
+subsubtests = inspect_container_simple
 
 [docker_cli/dockerinspect/inspect_container_simple]
 check_fields = ID,Image,Config
