+++ subtests/docker_cli/dockerinspect/dockerinspect.py
@@ -4,7 +4,6 @@ Test output of docker inspect command
 1. Create some docker containers
 2. Run docker inspect command on them
 3. Check output
-4. Compare output with values obtained in the container's config
 """
 
 # Okay to be less-strict for these cautions/warnings in subtests
@@ -22,9 +21,11 @@ import json
 import os
 
 class dockerinspect(SubSubtestCaller):
+
     config_section = 'docker_cli/dockerinspect'
 
 class inspect_base(SubSubtest):
+
     @staticmethod
     def verify_same_configs(subtest, source, comp, ignore_fields=[]):
         for i in range(len(comp)):
@@ -32,10 +33,11 @@ class inspect_base(SubSubtest):
                 if key in ignore_fields:
                     continue
                 subtest.failif(not (source[i][key] == comp[i][key]),
-                        "CLI output differs from container config: "
-                        "ID: %s, %s : %s != %s : %s" % (comp[i]['ID'][:12],
-                                                        key, source[i][key],
-                                                        key, comp[i][key]))
+                               "CLI output differs from container config: "
+                               "ID: %s, %s: %s != "
+                               "%s: %s" % (comp[i]['ID'][:12],
+                                           key, source[i][key],
+                                           key, comp[i][key]))
 
     @staticmethod
     def get_cid_from_name(subtest, name):
@@ -44,7 +46,11 @@ class inspect_base(SubSubtest):
         return next(x.long_id for x in containers if x.container_name == name)
 
     def parse_cli_output(self, output):
-        return json.loads(output)
+        try:
+            output_map = json.loads(output)
+        except:
+            self.failif(True, "Unable to parse inspect output.")
+        return output_map
 
     def find_container_path(self, cid, docker_root='/var/lib/docker/'):
         """
@@ -74,7 +80,7 @@ class inspect_base(SubSubtest):
             json_data = open(container_path + i)
             data = json.load(json_data)
             if i == 'hostconfig.json':
-                data = {u'HostConfig' : data}
+                data = {u'HostConfig': data}
             config_data = dict(config_data.items() + data.items())
             json_data.close()
         return config_data
@@ -106,18 +112,20 @@ class inspect_base(SubSubtest):
         nfdc.execute()
         return name
 
-    @staticmethod
-    def clean_containers(subtest, containers):
-        if subtest.config['remove_after_test']:
-            dkrcmd = DockerCmd(subtest.parent_subtest,
+    def cleanup(self):
+        super(inspect_base, self).cleanup()
+        if self.config['remove_after_test'] and self.sub_stuff['containers']:
+            dkrcmd = DockerCmd(self.parent_subtest,
                                'rm',
-                               containers)
+                               self.sub_stuff['containers'])
             dkrcmd.execute()
 
 class inspect_container_simple(inspect_base):
+
     def initialize(self):
         super(inspect_container_simple, self).initialize()
         self.sub_stuff['name'] = self.create_simple_container(self)
+        self.sub_stuff['containers'] = [self.sub_stuff['name']]
 
     def run_once(self):
         super(inspect_container_simple, self).run_once()
@@ -133,33 +141,3 @@ class inspect_container_simple(inspect_base):
             self.failif(field not in cli_output[0],
                         "Field: '%s' not found in output." % (field))
 
-    def cleanup(self):
-        super(inspect_container_simple, self).cleanup()
-        self.clean_containers(self, [self.sub_stuff['name']])
-
-class inspect_all(inspect_base):
-    def initialize(self):
-        super(inspect_all, self).initialize()
-        self.sub_stuff['name'] = self.create_simple_container(self)
-
-    def run_once(self):
-        super(inspect_all, self).run_once()
-        # find inputs to this
-        subargs = [self.sub_stuff['name']]
-        nfdc = NoFailDockerCmd(self.parent_subtest, "inspect", subargs)
-        self.sub_stuff['cmdresult'] = nfdc.execute()
-
-    def postprocess(self):
-        super(inspect_all, self).postprocess()
-        cli_output = self.parse_cli_output(self.sub_stuff['cmdresult'].stdout)
-        cid = self.get_cid_from_name(self, self.sub_stuff['name'])
-        config_map = self.get_config_maps([cid])
-        ifields = self.config['ignore_fields'].split(',')
-        self.verify_same_configs(self,
-                                 config_map,
-                                 cli_output,
-                                 ignore_fields=ifields)
-
-    def cleanup(self):
-        super(inspect_all, self).cleanup()
-        self.clean_containers(self, [self.sub_stuff['name']])
